move Entity components into its own folder
