+++ client/src/components/elements/index.js
@@ -45,13 +45,6 @@ export { default as BaseForm } from './BaseForm';
 export { default as BiotypeSelect } from './BiotypeSelect';
 export { default as Button } from './Button';
 export { default as DocumentTitle } from './DocumentTitle';
-export { default as EntityNotFound } from './EntityNotFound';
-export { default as EntityDirectoryButton } from './EntityDirectoryButton';
-export { default as EntityCreate } from './EntityCreate';
-export { default as EntityEdit } from './EntityEdit';
-export { default as EntityEditNew } from './EntityEditNew';
-export { default as EntityDirectory } from './EntityDirectory';
-export { default as EntityProfile } from './EntityProfile';
 export { default as ErrorBoundary } from './ErrorBoundary';
 export { default as SimpleAjax } from './SimpleAjax';
 export { default as NotFound } from './NotFound';
