+++ client/src/containers/Entity/index.js
@@ -1,0 +1,7 @@
+export { default as EntityNotFound } from './EntityNotFound';
+export { default as EntityDirectoryButton } from './EntityDirectoryButton';
+export { default as EntityCreate } from './EntityCreate';
+export { default as EntityEdit } from './EntityEdit';
+export { default as EntityEditNew } from './EntityEditNew';
+export { default as EntityDirectory } from './EntityDirectory';
+export { default as EntityProfile } from './EntityProfile';
